# specifications/architecture/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DigitalBurj Logistics OS API",
    "version": "1.0.0",
    "description": "Versioned REST contract. Money = decimal strings. Commands are idempotent via Idempotency-Key; mutable aggregates use If-Match."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "SERVICE_UNAVAILABLE",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "TENANT_REQUIRED",
              "NOT_FOUND",
              "VALIDATION_FAILED",
              "VERSION_CONFLICT",
              "IDEMPOTENCY_KEY_REUSED",
              "IDEMPOTENCY_KEY_REQUIRED",
              "INVALID_STATE_TRANSITION",
              "DUPLICATE",
              "ACCOUNTING_PERIOD_CLOSED",
              "PERIOD_NOT_FOUND",
              "INSUFFICIENT_STOCK",
              "STOCK_ON_HOLD",
              "RELEASE_EVIDENCE_REQUIRED",
              "SEPARATION_OF_DUTIES",
              "OVER_ALLOCATION",
              "JOB_NOT_CLOSABLE",
              "CURRENCY_MISMATCH",
              "DOCUMENT_NOT_CLEAN",
              "WEBHOOK_SIGNATURE_INVALID",
              "AI_TOOL_DENIED",
              "AI_BUDGET_EXCEEDED",
              "BOOKING_OUTCOME_UNKNOWN",
              "QUALIFICATION_REQUIRED",
              "RATE_LIMITED",
              "INTERNAL"
            ]
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            }
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ]
      }
    },
...
  },
  "paths": {
...
    "/api/v1/knowledge": {
      "get": {
        "operationId": "listKnowledge",
        "tags": [
          "Governance"
        ],
        "summary": "Search scoped knowledge revisions",
        "description": "Permission: work.view",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "maxLength": 200
            },
            "required": false,
            "name": "search",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "sop",
                "policy",
                "manual",
                "customer_sop",
                "carrier_instruction",
                "customs_guide",
                "training"
              ]
            },
            "required": false,
            "name": "category",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createKnowledge",
        "tags": [
          "Governance"
        ],
        "summary": "Create an immutable draft revision",
        "description": "Permission: work.manage",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "legalEntityId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "branchId": {
                    "type": "string",
                    "nullable": true,
                    "format": "uuid",
                    "default": null
                  },
                  "key": {
                    "type": "string",
                    "pattern": "^[a-z][a-z0-9_-]{1,80}$"
                  },
                  "title": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 200
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "sop",
                      "policy",
                      "manual",
                      "customer_sop",
                      "carrier_instruction",
                      "customs_guide",
                      "training"
                    ]
                  },
                  "content": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 20000
                  },
                  "changeNote": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 2000
                  }
                },
                "required": [
                  "legalEntityId",
                  "key",
                  "title",
                  "category",
                  "content",
                  "changeNote"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/{id}/publish": {
      "post": {
        "operationId": "publishKnowledge",
        "tags": [
          "Governance"
        ],
        "summary": "Independently publish a knowledge revision",
        "description": "Permission: work.configure",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reviewNote": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 2000
                  }
                },
                "required": [
                  "reviewNote"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/{id}/acknowledge": {
      "post": {
        "operationId": "acknowledgeKnowledge",
        "tags": [
          "Governance"
        ],
        "summary": "Acknowledge a specific published revision and content hash",
        "description": "Permission: work.view",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risks": {
      "get": {
        "operationId": "listRisks",
        "tags": [
          "Governance"
        ],
        "summary": "Scoped risk register ranked by residual score",
        "description": "Permission: quality.view",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createRisk",
        "tags": [
          "Governance"
        ],
        "summary": "Register owned and scored risk with mitigation",
        "description": "Permission: quality.manage",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "legalEntityId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "branchId": {
                    "type": "string",
                    "nullable": true,
                    "format": "uuid",
                    "default": null
                  },
                  "title": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 200
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "financial",
                      "operational",
                      "customer",
                      "carrier",
                      "compliance",
                      "security",
                      "legal"
                    ]
                  },
                  "description": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 20000
                  },
                  "likelihood": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "impact": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "mitigation": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 20000
                  },
                  "ownerMembershipId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "nextReviewOn": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  }
                },
                "required": [
                  "legalEntityId",
                  "title",
                  "category",
                  "description",
                  "likelihood",
                  "impact",
                  "mitigation",
                  "ownerMembershipId",
                  "nextReviewOn"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risks/{id}": {
      "get": {
        "operationId": "getRisk",
        "tags": [
          "Governance"
        ],
        "summary": "Risk and immutable review history",
        "description": "Permission: quality.view",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risks/{id}/reviews": {
      "post": {
        "operationId": "reviewRisk",
        "tags": [
          "Governance"
        ],
        "summary": "Independent risk review with closure evidence",
        "description": "Permission: quality.hold.release",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "If-Match",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "outcome": {
                    "type": "string",
                    "enum": [
                      "open",
                      "mitigating",
                      "closed"
                    ]
                  },
                  "residualLikelihood": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "residualImpact": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "reviewNote": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 4000
                  },
                  "mitigation": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 20000
                  },
                  "nextReviewOn": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "evidenceDocumentId": {
                    "type": "string",
                    "nullable": true,
                    "format": "uuid",
                    "default": null
                  }
                },
                "required": [
                  "outcome",
                  "residualLikelihood",
                  "residualImpact",
                  "reviewNote",
                  "mitigation",
                  "nextReviewOn"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}